Move InMemoryTimerInternals to runners-core
